pyproject.toml
[project]
name = "rubiks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "datasets>=4.1.0",
    "random-user-agent>=1.0.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "rubik-cube>=0.0.2",
    "tokenizers>=0.22.0",
    "transformers>=4.56.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
